+++ test/Sentry.Unity.Tests/UnityIl2CppEventExceptionProcessorTests.cs
@@ -6,6 +6,7 @@ public class UnityIl2CppEventExceptionProcessorTests
 {
     [Test]
     [TestCase(null, null)]
+    [TestCase("", null)]
     [TestCase("f30fef22-d93e-7f60-0000-000000000000", "f30fef22d93e7f60")]
     [TestCase("0c9249e5-e223-8bd5-0000-000000000000", "0c9249e5e2238bd5")]
     [TestCase("6f42afa0-45c8-86e6-2372-a02513d55560", "6f42afa045c886e62372a02513d55560")]